Fix vsock for Node.js and modernize interface
The Node vsock implementation was buggy when there were more than two
requests or if the response was large enough to require multiple writes
to the Vsock. This is because the vsock `write` was not actually
preventing the read (which was blocking) from beginning. So we could get
stuck in situations where the read had begun, and was blocking a write,
and no more requests would arrive.

This commit fixes the problem by acutally making the read/write requests
non-blocking using Node's Readable and Writeable stream interface.

Moreover, the commit changes the interface for Node.js apps to use the
more contemporary Promise-style interface, which allows both the runtime
and app to use `async/await` syntax for much simpler to reason about
implementations.