Build Java apps in Docker & move deps to apps
We should always build inside a Docker (or other) container to ensure
that we're using a _specific_ version of the compiler and libraries that
will be compatible with the runtime version.

We also want to put app-specific packages in app-specific folders. For
now, there is some duplication of the `json-simple` package across Java
apps, but we probably want to get rid of that dependency in the apps
anyway (e.g. by deserializing to a generic object and serializing any
compatible object for example using Google's gson library)