No-Runtime Runtime
A runtime and hello app example for non-runtime languages (C & Rust,
e.g.).

The "runtime" component is responsible for setting up the environment,
taking snapshots but does not interpose on VSock interactions between
the app and the VMM. Instead it fork/execs the app which, in turn, does
it's own vsock communication.

There is a simple pipe-based signaling protocol for the app to indicate, first,
that the app is ready for the runtime to take a diff-snapshot and then,
second, for the runtime to indicate to the app that it is done
diff-snapshotting and the app should begin it's communication over
vsock.